WIP: Re-work FetchPushDown into a two-step rewrite
Instead of creating a new subRelation which contains the `_fetchid` and
other `requiredForQuery` symbols, the relation is now mutated inplace
and the `postFetchOutputs` are returned as part of a `FetchDescription`.

The actual `fetchOutputs` can later be created using the
`FetchDescription`. This should make it possible to bubble-up the
`FetchDescription` in simple subqueries to defer to fetch.